command model

+++ src/com/design/observer/demo4/EvenObserver.java
@@ -6,7 +6,7 @@ import java.util.Observer;
 public class EvenObserver implements Observer {
 
     public void update(Observable o, Object arg) {
-       if (arg == NumsObservable.EVEN) {
+       if ((Integer)arg == NumsObservable.EVEN) {
            NumsObservable myObserable = (NumsObservable) o;
            System.out.println("EvenObserver:Data has changed to " + myObserable.getData());
        }
