+++ src/com/design/observer/demo4/ODDObserver.java
@@ -13,7 +13,7 @@ public class ODDObserver implements Observer {
 
     @Override
     public void update(Observable o, Object arg) {
-        if(arg == NumsObservable.ODD){
+        if((Integer)arg == NumsObservable.ODD){
             NumsObservable myObserable = (NumsObservable) o;
             System.out.println("OddObserver:Data has changed to " + myObserable.getData());
         }
